backend/requirements.txt
# ========================================
# Medible - "Med + Edible"
# Food-Drug Interaction Checker
# ========================================

# Flask & Extensions
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-migrate==4.0.5
flask-limiter==3.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP Requests (for OpenFDA & USDA APIs)
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Password Hashing & JWT
werkzeug==3.0.1
PyJWT==2.8.0

# Validation
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0

# Production Server
gunicorn==21.2.0